+++ src/components/generic/form/index.js
@@ -1,5 +1,5 @@
 import FormSection from './FormSection.vue';
 import ToggleButton from './ToggleButton.vue';
-import { BtnDelete, BtnSave } from './buttons';
+import { BtnDelete, BtnSave, BtnCancel } from './buttons';
 
-export { FormSection, ToggleButton, BtnDelete, BtnSave };
+export { FormSection, ToggleButton, BtnDelete, BtnSave, BtnCancel };