+++ src/views/ObjectiveHome.vue
@@ -2,11 +2,38 @@
   <page-layout v-if="activeObjective">
     <template #default>
       <header>
-        <span class="title-label">{{ $t('general.objective') }}</span>
-        <h2 class="title-1">{{ activeObjective.name }}</h2>
-        <p v-if="activeObjective.description" class="description">
-          {{ activeObjective.description }}
-        </p>
+        <span class="title-label">
+          {{ $t('general.objective') }}
+        </span>
+
+        <div class="objective__heading">
+          <h2 class="title-1">{{ activeObjective.name }}</h2>
+          <div class="objective__edit">
+            <pkt-button
+              v-tooltip="$t('objective.change')"
+              icon-name="edit"
+              :text="$t('objective.change')"
+              :variant="'icon-only'"
+              skin="tertiary"
+              @onClick="$emit('click', toggleDrawer('objective'))"
+            />
+          </div>
+        </div>
+        <div class="objective__description">
+          <p v-if="activeObjective.description" class="description">
+            {{ activeObjective.description }}
+          </p>
+          <div data-mode="dark" class="objective__add-key-res">
+            <pkt-button
+              v-tooltip="$t('btn.createKeyResult')"
+              :text="$t('btn.createKeyResult')"
+              skin="primary"
+              variant="icon-left"
+              icon-name="plus-sign"
+              @onClick="$emit('click', toggleDrawer('keyResult'))"
+            />
+          </div>
+        </div>
       </header>
 
       <section>
@@ -16,12 +43,6 @@
           :heading="$t('empty.noKeyResults.heading')"
           :body="$t('empty.noKeyResults.body')"
         >
-          <router-link
-            v-if="hasEditRights"
-            :to="{ name: 'ItemAdmin', query: { tab: 'okr' } }"
-          >
-            {{ $t('empty.noKeyResults.linkText') }}
-          </router-link>
         </empty-state>
 
         <div class="key-results__list">
@@ -55,12 +76,13 @@
 </template>
 
 <script>
-import { mapGetters, mapState } from 'vuex';
+import { mapGetters, mapState, mapMutations } from 'vuex';
 import routerGuard from '@/router/router-guards/objectiveHome';
 import WidgetWrapper from '@/components/widgets/WidgetWrapper.vue';
 import WidgetObjectiveDetails from '@/components/widgets/WidgetObjectiveDetails.vue';
 import WidgetWeights from '@/components/widgets/WidgetWeights.vue';
 import ProgressionChart from '@/components/ProgressionChart.vue';
+import { PktButton } from '@oslokommune/punkt-vue2';
 
 export default {
   name: 'ObjectiveHome',
@@ -72,6 +94,7 @@ export default {
     WidgetObjectiveDetails,
     ProgressionChart,
     EmptyState: () => import('@/components/EmptyState.vue'),
+    PktButton,
   },
 
   beforeRouteUpdate: routerGuard,
@@ -117,6 +140,27 @@ export default {
         );
       },
     },
+    keyResults: {
+      immediate: true,
+      handler() {
+        this.keyRes = this.keyResults.filter(
+          (keyRes) => keyRes.objective === `objectives/${this.activeObjective.id}`
+        );
+      },
+    },
+  },
+
+  methods: {
+    ...mapMutations(['TOGGLE_DRAWER']),
+    toggleDrawer(type) {
+      this.TOGGLE_DRAWER({
+        type,
+        show: 'true',
+        data: {
+          objective: this.activeObjective,
+        },
+      });
+    },
   },
 };
 </script>
@@ -134,12 +178,32 @@ export default {
   }
 }
 
+.key-result__new {
+  display: flex;
+  justify-content: flex-start;
+}
+
 .objective__heading {
   display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-between;
 }
 
 .objective__heading-text {
   max-width: 46rem;
   margin-right: 2.5rem;
+}
+
+.objective__description {
+  display: flex;
+  flex-direction: row;
+  align-items: self-end;
+  justify-content: space-between;
+  padding-bottom: 1rem;
+}
+
+.objective__add-key-res {
+  margin-left: auto;
 }
 </style>